Updated DB table generation to use decimal 18, 2 (to include actual decimal places). Added order by name to lists queries to alphabatize comboboxes and list boxes. Hide transfer accounts in Transaction manager unless transfer checkbox is checked. Other minor UI fixes.

+++ BudgetAp/DatabaseInsertsAndMods.cs
@@ -27,7 +27,7 @@ namespace BudgetAp
             newTransaction.TransactionType = transType;
             newTransaction.CategoryID = categoryID;
             newTransaction.VendorID = vendorID;
-            newTransaction.Amount = amount;
+            newTransaction.Amount = (decimal)amount;
             newTransaction.Description = description;
             transactionsTable.InsertOnSubmit(newTransaction);
         }
@@ -171,7 +171,7 @@ namespace BudgetAp
             transactionsModificationQuery.Single().TransactionType = transType;
             transactionsModificationQuery.Single().CategoryID = categoryID;
             transactionsModificationQuery.Single().VendorID = vendorID;
-            transactionsModificationQuery.Single().Amount = amount;
+            transactionsModificationQuery.Single().Amount = (decimal)amount;
             transactionsModificationQuery.Single().Description = description;
         }
